// src/indirizzoSelezionato/src/app/models/indirizzo.ts
export class Indirizzo {

  id: number | undefined | null;
  idAnagrafica: number | undefined | null;
  idDocumento: number | undefined | null;
  tipoIndirizzo: string | undefined | null;
  denominazioneAlternativa: string | undefined | null;
  nazione: string | undefined | null;
  indirizzo: string | undefined | null;
  provincia: string | undefined | null;
  comune: string | undefined | null;
  frazione: string | undefined | null;
  regione: string | undefined | null;
  cap: string | undefined | null;
  codSdi: string | undefined | null;
  telefono1: string | undefined | null;
  telefono2: string | undefined | null;
  cellulare: string | undefined | null;
  telefono3: string | undefined | null;
  fax: string | undefined | null;
  email: string | undefined | null;
  pec: string | undefined | null;
  sito: string | undefined | null;
  referente: string | undefined | null;
  recapitoReferente: string | undefined | null;
  notes: string | undefined | null;

}