// src/indirizzoSelezionato/src/app/models/comune.ts
export default class Comune {

  id: number | undefined | null;
  regione: string | undefined | null;
  provincia: string | undefined | null;
  citta: string | undefined | null;
  cap: string | undefined | null;
  codice: string | undefined | null;
  provinciaCompleta: string | undefined | null;
  codCom: string | undefined | null;
  codProv: string | undefined | null;
  codReg: string | undefined | null;
  idfRifIpt: string | undefined | null;

}